Spelling error corrected.
There is a spelling error in the first line of the readme section. The sentence is written as, " This is a Tensorflow implemention of VGG 16 and VGG 19 based on....."
The word 'implemention' is wrongly misspelled here.
The correct sentence could be written as, " This is a Tensorflow implementation of VGG 16 and VGG 19 based on....."
Thank you.